# backend/doc/openapi.yml
openapi: 3.0.3
info:
  title: Tournament API
  description: API for managing tournaments
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

security:
  - basicAuth: [ ]

paths:
  /api/tournament:
    get:
      tags:
        - Tournament
      summary: List all tournaments
      description: Retrieves a list of all tournaments
      operationId: listTournaments
      responses:
        '200':
          description: A list of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string

    post:
      tags:
        - Tournament
      summary: Create a new tournament
      description: Creates a new tournament with DRAFT status
      operationId: createTournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTournamentRequest'
      responses:
        '201':
          description: Tournament created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /api/tournament/{id}:
    get:
      tags:
        - Tournament
      summary: Get a tournament by ID
      description: Retrieves a tournament by its ID
      operationId: getTournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the tournament to retrieve
      responses:
        '200':
          description: Tournament found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '404':
          description: Tournament not found
          content:
            text/plain:
              schema:
                type: string
  /api/tournament/{id}/status:
    patch:
      tags:
        - Tournament
      summary: Update tournament status
      description: Updates the status of a tournament
      operationId: updateTournamentStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the tournament to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTournamentStatusRequest'
      responses:
        '200':
          description: Tournament status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        '400':
          description: Bad request or invalid status
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Tournament not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string

  /api/tournament/{id}/player:
    post:
      tags:
        - Player
      summary: Create a new player
      description: Create a new player and links them to the tournament
      operationId: createTournamentPlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTournamentPlayerRequest'
      responses:
        '200':
          description: Player created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
    get:
      tags:
        - Player
      summary: List all players
      description: List all players of the given tournament
      operationId: listTournamentPlayers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the tournament
      responses:
        '200':
          description: Player deleted successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'

  /api/tournament/{id}/player/{player_id}:
    delete:
      tags:
        - Player
      summary: Delete a player
      description: Delete the player with the given id from the tournament
      operationId: deleteTournamentPlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the tournament
        - name: player_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the player
      responses:
        '200':
          description: Player deleted successfully
          content:
            text/plain:
              schema:
                type: string

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    Tournament:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/TournamentStatus'
      required:
        - id
        - name
        - startDate
        - endDate
        - status
    TournamentStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - COMPLETED
        - CANCELLED
    Player:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tournamentId:
          type: string
    CreateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        allowUnderfilledGroups:
          type: boolean
        playerCount:
          type: integer
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/CreateTournamentRoundRequest'
      required:
        - name
        - startDate
        - endDate
        - allowUnderfilledGroups
        - playerCount
        - rounds
    CreateTournamentRoundRequest:
      type: object
      properties:
        name:
          type: string
        matchCount:
          type: integer
        playerAdvancementCount:
          type: integer
        groupSize:
          type: integer
        groupCount:
          type: integer
        concurrentGroupCount:
          type: integer
      required:
        - name
        - matchCount
        - playerAdvancementCount
        - groupSize
        - groupCount
        - concurrentGroupCount
    UpdateTournamentStatusRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TournamentStatus'
      required:
        - status
    CreateTournamentPlayerRequest:
      type: object
      properties:
        name:
          type: string